requirements.txt
numpy
pandas
openpyxl
odfpy
matplotlib
ipykernel
openai
flask
python-dotenv
opencv-python